pyproject.toml
[tool.poetry]
name = "irt2"
version = "1.1.0-rc0"
description = "IRT2 - Inductive Reasoning with Text"
authors = ["Felix Hamann <felix@hamann.xyz>"]
license = "MIT"
readme = "README.md"
packages = [{include = "irt2", from="src"}]

[tool.poetry.scripts]
irt2 = "irt2.cli:entry"

[tool.poetry.dependencies]
python = "^3.11"
ktz = "^0.4.0"
click = "^8.1.7"
matplotlib = "^3.8.4"
tabulate = "^0.9.0"
jupyter = "^1.0.0"
pretty-errors = "^1.2.25"
rich-click = "^1.7.4"
pyyaml = "^6.0.1"
networkx = "^3.3"

[tool.poetry.group.dev.dependencies]
pudb = "^2024.1"
ipython = "^8.23.0"
pyright = "^1.1.357"
black = "^24.3.0"
isort = "^5.13.2"
pytest = "^8.1.1"
pytest-pudb = "^0.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
